🚧 Extract function: `balloonCreator`
This makes it easier to see what's going on in the d3 code.

Note that this function is where all of the magic happens when it comes to appending a custom "template element" with d3. (Took me days to work this out, including a bunch of blog post reading, but also a bunch of toying around with the d3 source. Yipes!)